Refactor geocode service with plugin system
